added dependency injection for chatRoomFactory, StreamCollectionManager and ConnectivityManager
